Introduce BriteDatabase / BriteContentResolver types. Make SqlBrite an adapter for both.
