+++ sqlbrite-sample/src/main/java/com/example/sqlbrite/todo/db/DbModule.java
@@ -17,7 +17,7 @@ package com.example.sqlbrite.todo.db;
 
 import android.app.Application;
 import android.database.sqlite.SQLiteOpenHelper;
-import com.example.sqlbrite.todo.BuildConfig;
+import com.squareup.sqlbrite.BriteDatabase;
 import com.squareup.sqlbrite.SqlBrite;
 import dagger.Module;
 import dagger.Provides;
@@ -30,18 +30,15 @@ public final class DbModule {
     return new DbOpenHelper(application);
   }
 
-  @Provides @Singleton SqlBrite provideSqlBrite(SQLiteOpenHelper openHelper) {
-    SqlBrite db = SqlBrite.create(openHelper);
+  @Provides @Singleton SqlBrite provideSqlBrite() {
+    return SqlBrite.create(new SqlBrite.Logger() {
+      @Override public void log(String message) {
+        Timber.tag("Database").v(message);
+      }
+    });
+  }
 
-    if (BuildConfig.DEBUG) {
-      db.setLogger(new SqlBrite.Logger() {
-        @Override public void log(String message) {
-          Timber.tag("Database").v(message);
-        }
-      });
-      db.setLoggingEnabled(true);
-    }
-
-    return db;
+  @Provides @Singleton BriteDatabase provideDatabase(SqlBrite sqlBrite, SQLiteOpenHelper helper) {
+    return sqlBrite.wrapDatabaseHelper(helper);
   }
 }
